librbd: allow to override format in clone request

+++ src/librbd/image/CloneRequest.h
@@ -115,7 +115,7 @@ private:
   Context *m_on_finish;
 
   CephContext *m_cct;
-  uint32_t m_clone_format = 2;
+  uint64_t m_clone_format = 2;
   bool m_use_p_features;
   uint64_t m_features;
   map<string, bufferlist> m_pairs;
